Refactor: core-database 객체주입 모듈 위치 수정
